(More) correct date parsing for OSX.
If your locale is set to one that uses day-month ordering instead of month-day
when formatting dates, `ls` will show dates in that format and ls++ will get
very confused. This is a hack (ugly, but works) to remedy that -- it checks to
see if the day contains letters in it, and if it does, it assumes it's switched
around the day and the month.